Changed currency and income to float values from int

+++ Assets/_PROJECT/Scripts/Shop/ShopItemUnit.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 public class ShopItemUnit : ShopItem
 {
 
-    public int IncomeAddonFromSpawning;
+    public float IncomeAddonFromSpawning;
 
     public int XpAddonWhenBought;
